ci(api): environments
Rethought:  if get_environment() is a function in config that returns the Environment instance, we can be sure that each time it is loaded by some modules, it will catch the user defined env variable each time; and we can put that inside the qiboconnection.connection() constructor. Then, every time the user defines a new API(), the module will load with the currently defined env var that points to the correct environment.