feat(api): eliminated non-instance methods from `API` class
Segregated methods that did not use `self` from the `API` class. They have been moved to an `utils` class specially for `API`.

These functions are not base utilities in the sense that they depend on module-defined classes, as `JobResponse`. For this, they've been moved to an utils file for API.