feat(api): added `execute_and_return_results` function (#26)
* feat(api): added `execute_and_return_results` function

Added execute_and_return_results public method for executing jobs and synchronously wait for the results to be ready, and recover them automatically parsed.

* feat(api): eliminated non-instance methods from `API` class

Segregated methods that did not use `self` from the `API` class. They have been moved to an `utils` class specially for `API`.

These functions are not base utilities in the sense that they depend on module-defined classes, as `JobResponse`. For this, they've been moved to an utils file for API.

Co-authored-by: javi <javier.sabariego@qilimanjaro.tech>