feat(api): added missing required parameter to fav call
'user_id' is required for updating the favourite status of a favourite job.

The public service could deduce it from the logged in user, but as an admin could potentially decide to create Experiment-User relations or favourite-user relation, it is arguably more convenient for them to be explicitely provided by the client (and check they are trying to modify only their own user).