feat: qibo 179 save experiment and results (#30)
* fix(api): corrected urls used by save_experiment like functions

Due to copy paste issues, they were using jobs and circuits urls

* fix(api): misspelling in saved experiments url

* fix(api): adjusted response keywords

Adjusted SavedExperiments response classes. **Now we can successfully save and retrieve experiments from public** :D

* fix(api): adjusted response keywords

SavedExperimentListingResponse incorporates an id instead of a saved_experiment_id. Corrected

* fix(api): added favourite experiment creation

* feat(api): api.save_experiment() return behaviour

Now it behaves the same way that `api.execute()` does:  it updates the inner `connection._saved_experiment()` with the created ID, in case we want to do something, and it returns the id to the user.

* feat(api): api.list_saved_experiments() now follows pagination to retrieve full list

added logic to get full paginated list in api.list_saved_experiments. Added this logic also to `list_devices`, that also shared this problem

* feat(api): `SavedExperimentListing.dataframe` uses `id`s

SavedExperimentListing is now using the original item's id as the dataframe's index

* feat(api): `SavedExperimentListing.dataframe` uses `id`s

SavedExperimentListing is indexing by id, but kept as a column

* feat(api): api.list_saved_experiments() accepts `favourite` bool.

Modified api and connection so that we can choose to list all experiments or only the ones that are favourited to us

* feat(api): refactored paginated request iteration

Now, we `process_response()` each request before asking the next one, as if at any point we get an error, it may be missleading to try to continue

* fix(list_experimetns):  removed id column assignation as index

* fix(list_experimetns):  bad paginate iterator

iterating variable was not used, so it was looping forever...

* feat(saved_Experiments): added `qililab_version` parameter

Added `qililab_version` parameter to everything related to the saved experiments:  listing, retrieving and creating wise

* test(saved_Experiments): added saved_experiments related unittests

Added unittesting for `SavedExperiment` class, `SavedExperimentListing` and `SavedExperimentListingItem` classes, and for their related request/response typing classes

* feat(saved_Experiments): updated user_example.py

updated user_example python file with new qililab_version parameter con save_experiment functionality

* Revert "Revert "feat(api): added `save_experiment`, `list_saved_experiments`, and `get_experiments` functionalities""

This reverts commit 24c3a733e025b943427cd585c250b57ebc53a631.

* ci(api): solved mypy complaints

solved mypy complaints with api file

* ci(api): solved mypy complaints

solved mypy complaints with api file (again)

* ci(api): solved mypy complaints

solved mypy complaints with api file (again, last)

* fix(api): corrected numpy int used

numpy typing for int in our context is `np.int_`, which is the type compatible with python's `int`, as seen [here](https://numpy.org/devdocs/reference/arrays.scalars.html#numpy.int_).

Previously used `np.intp` is [probably just fine](https://numpy.org/devdocs/reference/arrays.scalars.html#numpy.intp), but less explicit.

* test(api): started testing api mocking connection

Tested `ping`, `list_devices`, `save_experiment`, `list_saved_experiments`,`get_saved_experiment`, `get_saved_experiments`,

* test(api): continued testing api mocking connection

Tested `select_device_id`, `select_device_ids`

* test(api): continued testing api mocking connection

Fix tests `select_device_id`, `select_device_ids`

* test(api): continued testing api mocking connection: mypy complaints

Solved mypy complaints on tests by asserting explicitely for some objects not to be nones

* feat(api): added methods for for favouritting an experiment

added methods for updating a saved experiment favourite association

* feat(api): added missing required parameter to fav call

'user_id' is required for updating the favourite status of a favourite job.

The public service could deduce it from the logged in user, but as an admin could potentially decide to create Experiment-User relations or favourite-user relation, it is arguably more convenient for them to be explicitely provided by the client (and check they are trying to modify only their own user).

* fix(api): success status for PUT

success status for a PUT request is 200, not 201. Corrected in request to savedexperiments endpoint

* fix(api): ensure dates are properly read

ensures that when the listing's dataframe reads the dates, they are read as datetimes and not as strings

* fix(api): adjusted tests

adjusted tests so that they chech for 200 status code on `PUT` to saved experiment endpoint

* feat(api): exposed `user_id`

Exposed user_id at the level of the api. Arguably useful, as it will for sure be used in more than one functionality to create user-related objects. This may be reverted if decided to.

* fix(api): removed unused imports and inapropiate ABC inheriting

Used of `ABC` as baseclass for `SavedExperiments`, `SavedExperimentListing` and `SavedExperimentListingItem` was not adequated, as those classes are inciated directly. Corrected.

* fix(api): user_id raises error if no user_id present

If we find a user that is not None but with a non identified user_id, we will raise an error.

* fix(api): corrected docstrings imprecissions

corrected copy-paste like mistakes

* fix(api): list comprehension style

reformatted a couple of response parsings into comprehension shape

* fix(api): comment misspelling

corrected misspelling of the word 'experiment'

* refactor(api): string constants extraction

repeated string constants have been extracted to a different file

* fix(api): list refactor issue

previous behaviour of `[response["items"] for response in responses]` mimics a `[].append(...)`-like behaviour, but we needed it to be `[].extend(...)`-like. A double comprehension fits that purpose: `[item for response in responses for item in response[REST.ITEMS]]`

* fix(api): corrected inner saved experiments and saved experiments lisitng overwritting

Now the saved experiment is only saved at the *upload* level:  when we create a new saved experiment, we keep the object locally referenced. User can access it using the `last_saved_experiment` endpoint.

The listing is never uploaded, so we keep the last *downloaded* one.

Co-authored-by: javi <javier.sabariego@qilimanjaro.tech>