feat(plotting): completed LivePlot testing
Added the required test for LivePlot class to be 100% covered.

Notes: Python unittest library has an `AsyncMock` class, that in theory should be automatically assigned to patched async functions instead of the usual `MagicMock`. That class can be naturally awaited, and has check functions for asserting it has been awaited with the proper objects.
For some reason, `unittest` does not interpretate `websockets.connect` function as async, so it forces it to be a usual mock which calls cannot be awaited.  This has been solved by making that mocks explicitely return futures.