Fix/websocket points lose (#39)
* WIP fix(tests): tried changing websocket library

i think the websocket error stuff is in qiboconnection's ws client, so lets try to change the library to the newwer asyncio one

* WIP fix(tests): continued using builtin websockets + asyncio queues

* fix(tests): compacting messages

as now we have a queue, we can ask ourselves whether we have accumulated several messages, and send a single packet containing them all. Done a first implementation of this

* WIP fix(tests): introduced timed connection reset

introduced a trigger to reset the connection after X seconds, that seemingly guarantees the server is kept alive without loosing point. Currently hardcoded to 5 seconds, to be set as environment

* feat(plotting): set data sending thread as a Daemon

Setting it as a daemon allows it to be automatically gracefully killed when the main process stops. We should still implement some function to allow the user to kill those threads, but for a first implementation, it seems ok to me.

* fix(tests): changed websocket module in tests

adapted tests to reflect the change in websocket module

* feat(plotting): completed LivePlots testing

Added the required test for LivePlots class to be 100% covered

* feat(plotting): completed LivePlot testing

Added the required test for LivePlot class to be 100% covered.

Notes: Python unittest library has an `AsyncMock` class, that in theory should be automatically assigned to patched async functions instead of the usual `MagicMock`. That class can be naturally awaited, and has check functions for asserting it has been awaited with the proper objects.
For some reason, `unittest` does not interpretate `websockets.connect` function as async, so it forces it to be a usual mock which calls cannot be awaited.  This has been solved by making that mocks explicitely return futures.

* fix(plotting): pylint complaints

* fix(plotting): pylint complaints

* test(plotting): async event loop refactor

tiny refactor of function to get current async event loop

* test(plotting): mypy complaints

* test(plotting): mypy complaints

* test(api): live-plots tests

* test(api): live-plot tests

* test(api): live-plot tests

---------

Co-authored-by: javi <javier.sabariego@qilimanjaro.tech>